app/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
langchain>=0.0.311
langchain-openai>=0.0.2
openai>=1.0.0
elevenlabs>=0.2.24
python-multipart>=0.0.6
starlette>=0.27.0
pydantic>=2.4.2
ffmpeg-python>=0.2.0
python-jose>=3.3.0
requests>=2.31.0